refactor: #1193 file-size ratchet 超過モジュールを分割
- project_authority.rs から identity プリミティブを project_identity.rs へ分離
- files.rs の画像 preview / 認可ゲートを files/image_preview.rs, files/root_gate.rs へ分離
- settings.rs の custom mascot 経路を settings_mascot.rs へ分離
- vibe_team_skill.rs / app/team_mcp.rs / team_hub/state/persistence.rs /
  project_authority.rs の inline test を child module へ移動 (既存 secure_install_tests 方式)
- 数行の authz gate 追加のみで分割余地がない shared.ts / git.rs / handoffs.rs /
  team_state.rs は baseline を更新 (縮小 8 ファイルの引き下げも反映)
