+++ src-tauri/src/commands/project_authority.rs
@@ -6,6 +6,7 @@
 
 use crate::commands::atomic_write::atomic_write_with_mode;
 use crate::commands::error::{CommandError, CommandResult};
+use crate::commands::project_identity::{canonical_root_key, PROJECT_AUTHORITY_SCHEMA_VERSION};
 use crate::state::AppState;
 use once_cell::sync::Lazy;
 use serde::{Deserialize, Serialize};
@@ -13,21 +14,11 @@ use std::path::{Path, PathBuf};
 use tauri::{AppHandle, Manager, State};
 use tokio::sync::Mutex;
 
-const PROJECT_AUTHORITY_SCHEMA_VERSION: u8 = 1;
+// identity プリミティブは `project_identity.rs` が所有する。外部モジュールは従来どおり
+// `project_authority::{ProjectRootIdentity, capture_identity, verify_identity}` を参照できる。
+pub use crate::commands::project_identity::{capture_identity, verify_identity, ProjectRootIdentity};
 
 static LEDGER_WRITE_LOCK: Lazy<Mutex<()>> = Lazy::new(|| Mutex::new(()));
-
-/// directory を path の表記ではなく filesystem object として識別する snapshot。
-///
-/// これにより、同じpathが別directoryへ置換された場合や symlink の参照先が変わった場合を
-/// fail-closed に検知する。各フィールドは ledger の wire format であり、renderer 入力は使わない。
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ProjectRootIdentity {
-    pub version: u8,
-    pub canonical_root: String,
-    pub platform_file_id: String,
-}
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -49,121 +40,6 @@ pub struct PickedProjectFile {
 
 fn authority_path() -> PathBuf {
     crate::util::config_paths::project_authority_path()
-}
-
-fn canonical_root_key(path: &Path) -> String {
-    let normalized = path.to_string_lossy().replace('\\', "/");
-    let stripped = normalized.trim_end_matches('/');
-    let normalized = if stripped.is_empty() && normalized.starts_with('/') {
-        "/"
-    } else {
-        stripped
-    };
-    if cfg!(windows) {
-        normalized.to_lowercase()
-    } else {
-        normalized.to_string()
-    }
-}
-
-#[cfg(unix)]
-fn platform_file_id(path: &Path) -> Result<String, CommandError> {
-    use std::os::unix::fs::MetadataExt;
-
-    let metadata = std::fs::metadata(path)
-        .map_err(|error| CommandError::authz(format!("metadata project root failed: {error}")))?;
-    Ok(format!("unix:{}:{}", metadata.dev(), metadata.ino()))
-}
-
-#[cfg(windows)]
-fn platform_file_id(path: &Path) -> Result<String, CommandError> {
-    use std::os::windows::{fs::OpenOptionsExt, io::AsRawHandle};
-    use windows_sys::Win32::Storage::FileSystem::{
-        GetFileInformationByHandle, BY_HANDLE_FILE_INFORMATION, FILE_FLAG_BACKUP_SEMANTICS,
-    };
-
-    let file = std::fs::OpenOptions::new()
-        .read(true)
-        .custom_flags(FILE_FLAG_BACKUP_SEMANTICS)
-        .open(path)
-        .map_err(|error| CommandError::authz(format!("open project root failed: {error}")))?;
-    let mut info: BY_HANDLE_FILE_INFORMATION = unsafe { std::mem::zeroed() };
-    let ok = unsafe {
-        GetFileInformationByHandle(file.as_raw_handle() as _, std::ptr::addr_of_mut!(info))
-    };
-    if ok == 0 {
-        return Err(CommandError::authz(format!(
-            "GetFileInformationByHandle project root failed: {}",
-            std::io::Error::last_os_error()
-        )));
-    }
-    let file_index = (u64::from(info.nFileIndexHigh) << 32) | u64::from(info.nFileIndexLow);
-    Ok(format!(
-        "windows:{}:{file_index}",
-        info.dwVolumeSerialNumber
-    ))
-}
-
-#[cfg(not(any(unix, windows)))]
-fn platform_file_id(_path: &Path) -> Result<String, CommandError> {
-    Err(CommandError::authz(
-        "project filesystem identity is unsupported on this platform",
-    ))
-}
-
-/// canonicalize と identity取得の間に root が差し替えられていないことを二重snapshotで確認する。
-/// directory が継続的に変化している場合は推測せず fail-closed にする。
-fn capture_identity_blocking(candidate: PathBuf) -> CommandResult<ProjectRootIdentity> {
-    if !crate::commands::fs_watch::is_safe_watch_root(&candidate) {
-        return Err(CommandError::validation(
-            "project root rejected by safety check (system / home / non-existent dir)",
-        ));
-    }
-
-    for _ in 0..3 {
-        let first = std::fs::canonicalize(&candidate).map_err(|error| {
-            CommandError::authz(format!("canonicalize project root failed: {error}"))
-        })?;
-        let first_id = platform_file_id(&first)?;
-        let second = std::fs::canonicalize(&candidate).map_err(|error| {
-            CommandError::authz(format!("canonicalize project root failed: {error}"))
-        })?;
-        let second_id = platform_file_id(&second)?;
-        if first == second && first_id == second_id {
-            return Ok(ProjectRootIdentity {
-                version: PROJECT_AUTHORITY_SCHEMA_VERSION,
-                canonical_root: canonical_root_key(&second),
-                platform_file_id: second_id,
-            });
-        }
-    }
-
-    Err(CommandError::authz(
-        "project root changed while its identity was being verified",
-    ))
-}
-
-pub async fn capture_identity(candidate: impl Into<PathBuf>) -> CommandResult<ProjectRootIdentity> {
-    let candidate = candidate.into();
-    tokio::task::spawn_blocking(move || capture_identity_blocking(candidate))
-        .await
-        .map_err(|error| CommandError::internal(format!("project identity task failed: {error}")))?
-}
-
-pub async fn verify_identity(identity: &ProjectRootIdentity) -> CommandResult<()> {
-    if identity.version != PROJECT_AUTHORITY_SCHEMA_VERSION
-        || identity.canonical_root.trim().is_empty()
-        || identity.platform_file_id.trim().is_empty()
-    {
-        return Err(CommandError::authz("invalid stored project root authority"));
-    }
-    let current = capture_identity(identity.canonical_root.clone()).await?;
-    if &current != identity {
-        return Err(CommandError::authz(
-            "project root identity no longer matches its native approval",
-        ));
-    }
-    Ok(())
 }
 
 async fn load_ledger_from(path: &Path) -> CommandResult<ProjectAuthorityLedger> {
@@ -476,53 +352,5 @@ pub async fn is_authorized_active_root(requested_canonical: &Path) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::tempdir;
-
-    #[tokio::test]
-    async fn ledger_roundtrip_is_private_and_versioned() {
-        let dir = tempdir().unwrap();
-        let path = dir.path().join("project-authority.json");
-        let root = dir.path().join("project");
-        tokio::fs::create_dir_all(&root).await.unwrap();
-        let identity = capture_identity(root).await.unwrap();
-        let ledger = ProjectAuthorityLedger {
-            schema_version: PROJECT_AUTHORITY_SCHEMA_VERSION,
-            active: Some(identity.clone()),
-            workspace_roots: vec![identity],
-        };
-        write_ledger_to(&path, &ledger).await.unwrap();
-        assert_eq!(load_ledger_from(&path).await.unwrap().active, ledger.active);
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            assert_eq!(
-                std::fs::metadata(&path).unwrap().permissions().mode() & 0o777,
-                0o600
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn missing_ledger_does_not_migrate_renderer_settings_candidates() {
-        let dir = tempdir().unwrap();
-        let missing = dir.path().join("project-authority.json");
-        let ledger = load_ledger_from(&missing).await.unwrap();
-        assert!(ledger.active.is_none());
-        assert!(ledger.workspace_roots.is_empty());
-        assert_eq!(ledger.schema_version, PROJECT_AUTHORITY_SCHEMA_VERSION);
-    }
-
-    #[tokio::test]
-    async fn replacement_at_same_path_invalidates_identity() {
-        let dir = tempdir().unwrap();
-        let root = dir.path().join("project");
-        let old = dir.path().join("old-project");
-        tokio::fs::create_dir_all(&root).await.unwrap();
-        let identity = capture_identity(&root).await.unwrap();
-        tokio::fs::rename(&root, &old).await.unwrap();
-        tokio::fs::create_dir_all(&root).await.unwrap();
-        assert!(verify_identity(&identity).await.is_err());
-    }
-}
+#[path = "project_authority/tests.rs"]
+mod tests;
